SS-7800: preparation for selection parameter API
